feat: implement WebSocket protocol handlers and schema merging utilities
